move invalid data exception to utils
