+++ common-util/src/main/java/com/vkondrav/ram/util/InvalidDataException.kt
@@ -1,3 +1,3 @@
-package com.vkondrav.ram.graphql.error
+package com.vkondrav.ram.util
 
 data class InvalidDataException(override val message: String) : Throwable(message)
