+++ domain/src/main/java/com/vkondrav/ram/domain/RamEpisodeDetails.kt
@@ -1,6 +1,6 @@
 package com.vkondrav.ram.domain
 
-import com.vkondrav.ram.graphql.error.InvalidDataException
+import com.vkondrav.ram.util.InvalidDataException
 import com.vkondrav.ram.graphql.generated.EpisodeDetailsQuery
 import kotlinx.coroutines.flow.Flow
 import timber.log.Timber
