+++ domain/src/main/java/com/vkondrav/ram/domain/RamLocation.kt
@@ -1,6 +1,6 @@
 package com.vkondrav.ram.domain
 
-import com.vkondrav.ram.graphql.error.InvalidDataException
+import com.vkondrav.ram.util.InvalidDataException
 import com.vkondrav.ram.graphql.generated.fragment.LocationFragment
 import com.vkondrav.ram.room.FavoriteLocation
 import kotlinx.coroutines.flow.Flow
