+++ domain/src/test/java/com/vkondrav/ram/domain/RamCharacterTest.kt
@@ -1,7 +1,7 @@
 package com.vkondrav.ram.domain
 
 import app.cash.turbine.test
-import com.vkondrav.ram.graphql.error.InvalidDataException
+import com.vkondrav.ram.util.InvalidDataException
 import com.vkondrav.ram.graphql.generated.fragment.CharacterFragment
 import com.vkondrav.ram.room.FavoriteCharacter
 import com.vkondrav.ram.test.BaseTest
